requirements.txt
django-filter
django
djangorestframework
python-decouple
psycopg2-binary
dj_config_url
drf_yasg
djangorestframework_simplejwt
